pyproject.toml
[project]
name = "cv-tailorer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fpdf2>=2.8.5",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-ollama>=1.0.0",
    "ollama>=0.6.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
